Simple Coloured map from height
A very basic version of a coloured map, representing biomes has been implemented. It does not work fully yet as vertex shading needs to be implemented so the entire chunk goes the colour of the highest vertex for the time being. The colour and heights of each layer colour can be decided in the Unity inspector.